requirements.txt
torch
#torchsummary
torchvision >= 0.3
# Optionally install via 'pip install git+https://github.com/szagoruyko/pytorchviz'
# pytorchviz
#pytorch-ignite
pycocotools # For Windows, install via 'pip install git+https://github.com/philferriere/cocoapi.git#subdirectory=PythonAPI'

codecov
pytest
pytest-cov
matplotlib
scipy
pillow
tqdm
omrdatasettools >= 0.19
mung